Add transparent pagination to query tabs, also limit result size.
Previously a query with 1M result rows would evaluate all 1M rows and
only afterwards trim to the number displayed per page (1000 by default).
This fixes the performance issue by applying a LIMIT to a wrapped
version of the query, and perhaps redundantly, also using the fetchmany()
API instead of fetchall().

Since we were in there, we added transparent pagination as well, and a
new configuration option to set the results-per-page on the query tabs.

Fixes #153